sorting: Prepare for multiple sorting options
Keep a global state for each sortable list instead of acting only on the
input of the controller that was activated.  This will allow multiple
sorting options to be used together.

Also find the related controls and container once during initialization
and store them with the state so we don't need to search for them every
time the sorting is changed.

Additionally, initialize the sorted state directly to the cookie value
instead of first sorting alphabetically and then resorting according to
the cookie.