fix: repopulate expansion items and remove debugging
A refactor resulted in us trying to foreach over
the Map of EXPANSION_LINKS in a way that resulted
in no items being yielded and no expansion links
being added. To resolve this, we changed the Map
to a tuple array. This commit also tears out all
the debugging stuff and simplifies the
construction of the expansions list.